requirements.txt
Flask>=3.0.0,<4.0.0
pypdf>=5.0.0,<6.0.0
psycopg2-binary>=2.9,<3.0
requests>=2.31,<3.0
gunicorn>=21.0,<22.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
pytest-cov>=5.0,<6.0
responses>=0.25,<1.0
